pyproject.toml
[project]
name = "pixiu-claude"
version = "0.1.0"
description = "貔貅 - 智能基金交易分析系统"
requires-python = ">=3.12"
dependencies = [
    "akshare>=1.18.25",
    "anthropic>=0.52.0",
    "google-genai>=1.0.0",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pydantic>=2.0",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "schedule>=1.2.2",
    "scipy>=1.17.0",
    "tabulate>=0.9.0",
    "lark-oapi>=1.3.0",
]

[project.scripts]
pixiu = "src.main:main"
pixiu-bot = "src.bot.app:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
